refactor : domain eentity 구조 변경
