+++ src/main/java/org/tuna/zoopzoop/backend/domain/archive/archive/entity/Archive.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.archive.entity;
+package org.tuna.zoopzoop.backend.domain.archive.archive.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,6 +7,7 @@ import jakarta.persistence.Enumerated;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.tuna.zoopzoop.backend.domain.archive.archive.enums.ArchiveType;
 import org.tuna.zoopzoop.backend.global.jpa.entity.BaseEntity;
 
 @Getter
