+++ src/main/java/org/tuna/zoopzoop/backend/domain/archive/folder/entity/Folder.java
@@ -1,9 +1,10 @@
-package org.tuna.zoopzoop.backend.domain.archive.entity;
+package org.tuna.zoopzoop.backend.domain.archive.folder.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.tuna.zoopzoop.backend.domain.archive.archive.entity.Archive;
 import org.tuna.zoopzoop.backend.global.jpa.entity.BaseEntity;
 
 @Getter
@@ -19,4 +20,8 @@ public class Folder extends BaseEntity {
     //폴더 이름
     @Column(nullable = false)
     private String name;
+
+    //디폴트 폴더 여부
+    @Column(nullable = false)
+    private boolean isDefault = false;
 }
