+++ src/main/java/org/tuna/zoopzoop/backend/domain/datasource/entity/DataSource.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.tuna.zoopzoop.backend.domain.archive.entity.Folder;
+import org.tuna.zoopzoop.backend.domain.archive.folder.entity.Folder;
 import org.tuna.zoopzoop.backend.global.jpa.entity.BaseEntity;
 
 import java.time.LocalDateTime;
@@ -32,7 +32,7 @@ public class DataSource extends BaseEntity {
     //소스 데이터의 작성일자
     //DB 저장용 createdDate와 다름.
     @Column(nullable = false)
-    private LocalDateTime dateCreatedDate;
+    private LocalDateTime dataCreatedDate;
 
     //소스 데이터 URL
     @Column(nullable = false)
@@ -42,6 +42,11 @@ public class DataSource extends BaseEntity {
     @Column
     private String thumbnailUrl;
 
+    // 태그 목록
     @OneToMany(mappedBy = "dataSource", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Tag> tags = new ArrayList<>();
+
+    // 활성화 여부
+    @Column(nullable = false)
+    private boolean isActive = true;
 }
