+++ src/main/java/org/tuna/zoopzoop/backend/domain/datasource/entity/Tag.java
@@ -18,5 +18,5 @@ public class Tag extends BaseEntity {
 
     //태그명
     @Column(nullable = false)
-    private String tag;
+    private String tagName;
 }
